requirements.txt
dnspython
mysql-connector-python
